refactor(files): refactor name package model to package entity
