+++ src/main/kotlin/ecommerce/controller/CartAdminController.kt
@@ -3,7 +3,7 @@ package ecommerce.controller
 import ecommerce.auth.annotation.LoginMember
 import ecommerce.dto.MemberStatsResponse
 import ecommerce.dto.ProductStatResponse
-import ecommerce.model.Member
+import ecommerce.entity.Member
 import ecommerce.service.CartService
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
