+++ src/main/kotlin/ecommerce/entity/Product.kt
@@ -1,4 +1,4 @@
-package ecommerce.model
+package ecommerce.entity
 
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
