+++ src/main/kotlin/ecommerce/repository/ProductJpaRepository.kt
@@ -1,6 +1,6 @@
 package ecommerce.repository
 
-import ecommerce.model.Product
+import ecommerce.entity.Product
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
