+++ src/test/kotlin/ecommerce/repository/MemberJpaRepositoryTest.kt
@@ -1,7 +1,6 @@
 package ecommerce.repository
 
-import ecommerce.model.Member
-import ecommerce.model.Product
+import ecommerce.entity.Member
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
