+++ src/test/kotlin/ecommerce/repository/ProductJpaRepositoryTest.kt
@@ -1,6 +1,6 @@
 package ecommerce.repository
 
-import ecommerce.model.Product
+import ecommerce.entity.Product
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.springframework.beans.factory.annotation.Autowired
